+++ site/src/main.tsx
@@ -1,0 +1,19 @@
+import "eeve/styles.css";
+import "./site.css";
+
+import * as React from "react";
+import { createRoot } from "react-dom/client";
+
+import { App } from "./App";
+
+const rootElement = document.getElementById("root");
+
+if (!rootElement) {
+  throw new Error("Root element was not found.");
+}
+
+createRoot(rootElement).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);